optimized api calls, changed names of the component, fixed padding in the topic form
